groundwork for tabbed swipe view
