+++ app/src/main/java/fho/kdvs/favorite/broadcast/FavoriteBroadcastFragment.kt
@@ -1,0 +1,6 @@
+package fho.kdvs.favorite.broadcast
+
+import fho.kdvs.global.BaseFragment
+
+class FavoriteBroadcastFragment: BaseFragment() {
+}
