+++ app/src/main/java/fho/kdvs/track/BroadcastTrackDetailsViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
 import androidx.navigation.NavController
 import fho.kdvs.broadcast.BroadcastRepository
-import fho.kdvs.favorite.FavoriteRepository
+import fho.kdvs.favorite.track.FavoriteTrackRepository
 import fho.kdvs.global.database.BroadcastEntity
 import fho.kdvs.global.database.FavoriteTrackEntity
 import fho.kdvs.global.database.ShowEntity
@@ -16,7 +16,7 @@ import javax.inject.Inject
 class BroadcastTrackDetailsViewModel @Inject constructor(
     val trackRepository: TrackRepository,
     private val broadcastRepository: BroadcastRepository,
-    private val favoriteRepository: FavoriteRepository,
+    private val favoriteRepository: FavoriteTrackRepository,
     application: Application
 ) : AndroidViewModel(application) {
